Bugfix: limit created families to prevent crash
Update create_nobility() to only create families if we have names
available to create them with.
Update create_nobility() to return feedback dict with requested, created
families and whether it was cut off or not.
Update realm.html to display number of created families, number of
requested families in red if input was too large.